pyproject.toml
[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"


[project]
name = "mlkit"
version = "0.1.0"
description = "AI-maintained swiss army knife for scientific scripts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
"ase> = 3.26.0",
"matplotlib> = 3.10.8",
"numpy> = 2.2.6",
"pandas> = 2.3.3",
"pymatgen> = 2025.10.7",
"pyyaml> = 6.0.3",
"ruamel.yaml> = 0.18.10",
"seekpath> = 2.1.0",
"typer> = 0.20.1"
]


[project.scripts]
mlkit = "mlkit.main:app"


[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
